+++ tests/migration/__init__.py
@@ -1,3 +1,5 @@
+from convert_to_new_inheritance_model import *
+from refrencefield_dbref_to_object_id import *
 from turn_off_inheritance import *
 
 if __name__ == '__main__':